+++ src/persistence.rs
@@ -200,70 +200,6 @@ impl Persistence {
         Ok(associations)
     }
 
-    /// Get the number of associations in a namespace.
-    pub async fn association_count(&self, ns: &str) -> Result<usize> {
-        let _permit = self.acquire_remote_slot().await?;
-        let conn = self.connect().await?;
-
-        let mut rows = conn
-            .query(
-                "SELECT COUNT(*) FROM csm_associations WHERE namespace = ?1",
-                params![ns],
-            )
-            .await
-            .map_err(|e| MemoryError::database(format!("Failed to count associations: {e}")))?;
-
-        if let Some(row) = rows
-            .next()
-            .await
-            .map_err(|e| MemoryError::database(format!("Failed to fetch row: {e}")))?
-        {
-            let count: i64 = row
-                .get(0)
-                .map_err(|e| MemoryError::database(format!("Failed to get count: {e}")))?;
-            Ok(count as usize)
-        } else {
-            Ok(0)
-        }
-    }
-
-    pub async fn for_each_association_scoped<F, Fut>(&self, ns: &str, mut f: F) -> Result<()>
-    where
-        F: FnMut(String, String, f32) -> Fut,
-        Fut: std::future::Future<Output = Result<()>>,
-    {
-        let _permit = self.acquire_remote_slot().await?;
-        let conn = self.connect().await?;
-
-        let mut rows = conn
-            .query(
-                "SELECT from_id, to_id, strength FROM csm_associations WHERE namespace = ?1",
-                params![ns],
-            )
-            .await
-            .map_err(|e| MemoryError::database(format!("Failed to stream associations: {e}")))?;
-
-        while let Some(row) = rows
-            .next()
-            .await
-            .map_err(|e| MemoryError::database(format!("Failed to fetch association row: {e}")))?
-        {
-            let from_id: String = row
-                .get(0)
-                .map_err(|e| MemoryError::database(format!("Failed to get from_id: {e}")))?;
-            let to_id: String = row
-                .get(1)
-                .map_err(|e| MemoryError::database(format!("Failed to get to_id: {e}")))?;
-            let strength: f64 = row
-                .get(2)
-                .map_err(|e| MemoryError::database(format!("Failed to get strength: {e}")))?;
-
-            f(from_id, to_id, strength as f32).await?;
-        }
-
-        Ok(())
-    }
-
     /// Perform database checkpoint (optimize)
     pub async fn checkpoint(&self) -> Result<()> {
         let _permit = self.acquire_remote_slot().await?;
