+++ src/singularity.rs
@@ -413,11 +413,6 @@ impl Singularity {
         self.get_namespace(ns).map_or(0, |n| n.concepts.len())
     }
 
-    pub fn association_count(&self, ns: &str) -> usize {
-        self.get_namespace(ns)
-            .map_or(0, |n| n.associations.values().map(|m| m.len()).sum())
-    }
-
     pub fn is_empty(&self, ns: &str) -> bool {
         self.get_namespace(ns).is_none_or(|n| n.concepts.is_empty())
     }
