+++ tests/ann_filter_bug.rs
@@ -57,8 +57,6 @@ async fn test_hnsw_filter_bug_prevention() {
     // Should only contain "match", not "no-match"
     assert!(results.iter().any(|(id, _)| id == "match"));
     assert!(!results.iter().any(|(id, _)| id == "no-match"));
-
-    drop(framework);
 }
 
 #[cfg(feature = "ann-hnsw")]
@@ -92,8 +90,6 @@ async fn test_hnsw_empty_filter_results() {
         "Results should be empty when filter matches nothing, but got {:?}",
         results
     );
-
-    drop(framework);
 }
 
 #[cfg(feature = "ann-lsh")]
@@ -144,6 +140,4 @@ async fn test_lsh_filter_bug_prevention() {
 
     assert!(results.iter().any(|(id, _)| id == "match"));
     assert!(!results.iter().any(|(id, _)| id == "no-match"));
-
-    drop(framework);
 }
